🏗 vue-dot: Refactor folder structure
