+++ packages/vue-dot/src/elements/CopyBtn/config.ts
@@ -1,0 +1,13 @@
+export default {
+	btn: {
+		icon: true,
+		class: 'grey--text text--darken-3 copy-tooltip-activator'
+	},
+	tooltip: {
+		/** By default, the tooltip is placed on the right */
+		right: true,
+		openOnClick: true,
+		openOnHover: false,
+		closeDelay: 2500
+	}
+};
