+++ packages/vue-dot/src/elements/CopyBtn/index.ts
@@ -1,0 +1,3 @@
+import CopyBtn from './CopyBtn.vue';
+
+export default CopyBtn;
