+++ packages/vue-dot/src/elements/CustomIcon/CustomIcon.vue
@@ -38,7 +38,7 @@
 	import Vue from 'vue';
 	import Component from 'vue-class-component';
 
-	import { Icons, Theme } from '../../types';
+	import { Icons, Theme } from '../../../types';
 
 	const Props = Vue.extend({
 		props: {
