+++ packages/vue-dot/src/elements/CustomIcon/index.ts
@@ -1,0 +1,3 @@
+import CustomIcon from './CustomIcon.vue';
+
+export default CustomIcon;
