+++ packages/vue-dot/src/elements/DataList/index.ts
@@ -1,0 +1,3 @@
+import DataList from './DataList.vue';
+
+export default DataList;
