+++ packages/vue-dot/src/elements/index.ts
@@ -1,0 +1,9 @@
+import CopyBtn from './CopyBtn';
+import CustomIcon from './CustomIcon';
+import DataList from './DataList';
+
+export default {
+	CopyBtn,
+	CustomIcon,
+	DataList
+};
