+++ packages/vue-dot/src/patterns/DatePicker/index.ts
@@ -1,0 +1,3 @@
+import DatePicker from './DatePicker.vue';
+
+export default DatePicker;
