+++ packages/vue-dot/src/patterns/FileList/config.ts
@@ -1,0 +1,21 @@
+export default {
+	listItemAvatarIcon: {
+		size: 24
+	},
+	layout: {
+		justifyEnd: true
+	},
+	divider: {
+		inset: true
+	},
+	retryBtn: {
+		icon: true
+	},
+	viewFileBtn: {
+		icon: true,
+		class: 'mr-2'
+	},
+	deleteFileBtn: {
+		icon: true
+	}
+};
