+++ packages/vue-dot/src/patterns/FileList/index.ts
@@ -1,0 +1,3 @@
+import FileList from './FileList.vue';
+
+export default FileList;
