+++ packages/vue-dot/src/patterns/FileUpload/index.ts
@@ -1,0 +1,3 @@
+import FileUpload from './FileUpload.vue';
+
+export default FileUpload;
