+++ packages/vue-dot/src/patterns/LangBtn/config.ts
@@ -1,0 +1,14 @@
+export default {
+	menu: {
+		offsetY: true
+	},
+	btn: {
+		color: 'currentColor',
+		depressed: true,
+		outlined: true,
+		ripple: true
+	},
+	icon: {
+		class: 'ml-1'
+	}
+};
