+++ packages/vue-dot/src/patterns/LangBtn/index.ts
@@ -1,0 +1,3 @@
+import LangBtn from './LangBtn.vue';
+
+export default LangBtn;
