+++ packages/vue-dot/src/patterns/NotificationBar/config.ts
@@ -1,0 +1,15 @@
+export default {
+	snackBar: {
+		value: true,
+		timeout: 0,
+		top: true
+	},
+	icon: {
+		color: 'white',
+		class: 'mr-2'
+	},
+	btn: {
+		color: 'white',
+		text: true
+	}
+};
