+++ packages/vue-dot/src/patterns/NotificationBar/index.ts
@@ -1,0 +1,3 @@
+import NotificationBar from './NotificationBar.vue';
+
+export default NotificationBar;
