+++ packages/vue-dot/src/patterns/PaginatedTable/PaginatedTable.vue
