+++ packages/vue-dot/src/patterns/PaginatedTable/index.ts
@@ -1,0 +1,3 @@
+import PaginatedTable from './PaginatedTable.vue';
+
+export default PaginatedTable;
