+++ packages/vue-dot/src/patterns/UploadWorkflow/config.ts
@@ -1,0 +1,29 @@
+export default {
+	fileUpload: {
+		class: 'mt-6'
+	},
+	dialog: {
+		persistent: true,
+		width: '550'
+	},
+	card: {
+		class: 'pa-4'
+	},
+	select: {
+		outlined: true,
+		validateOnBlur: true,
+		label: 'Nature du fichier'
+	},
+	layout: {
+		wrap: true,
+		class: 'mt-2'
+	},
+	cancelBtn: {
+		text: true,
+		class: 'mr-4',
+		color: 'accent'
+	},
+	confirmBtn: {
+		color: 'accent'
+	}
+};
