+++ packages/vue-dot/src/patterns/UploadWorkflow/index.ts
@@ -1,0 +1,3 @@
+import UploadWorkflow from './UploadWorkflow.vue';
+
+export default UploadWorkflow;
