+++ packages/vue-dot/src/patterns/index.ts
@@ -1,0 +1,17 @@
+import DatePicker from './DatePicker';
+import FileList from './FileList';
+import FileUpload from './FileUpload';
+import LangBtn from './LangBtn';
+import NotificationBar from './NotificationBar';
+import PaginatedTable from './PaginatedTable';
+import UploadWorkflow from './UploadWorkflow';
+
+export default {
+	DatePicker,
+	FileList,
+	FileUpload,
+	LangBtn,
+	NotificationBar,
+	PaginatedTable,
+	UploadWorkflow
+};
